Fix Pylance
Pylance couldn't find python libraries with the shim_scripts.  Changed
tactics to linking .venv to the pipenv environment.

The down-side of this is that `dev-scripts/setup-pipenv-link` will need
to be re-run anytime you `pipenv --rm && pipenv install`.  I wasn't able to find
any way to make pipenv automatically trigger that.